pyproject.toml
[tool.poetry]
name = "digichamp"
version = "0.1.0"
description = ""
authors = ["Divina Mbel <divinamanye@gmail.com>"]
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.10.12"
django = "^4.2"
djangorestframework = "^3.14"
djangorestframework-simplejwt = "^5.3"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2"
google-generativeai = "^0.3"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
